Working on new huddle, issues with huddle buttons

+++ Study Huddle/NSDateFormatter+RelativeString.m
@@ -24,12 +24,12 @@
     if ([dateDiff second] < 60) {
         [result appendString:@"less than 1 minute ago"];
     } else if ([dateDiff second] < SECONDS_PER_HOUR) {
-        [result appendFormat:@"%d minute%@ ago",
+        [result appendFormat:@"%ld minute%@ ago",
          ([dateDiff second] / 60) + 1,
          ([dateDiff minute] < 2) ? @"" : @"s"];
     } else if ([dateDiff second] < (SECONDS_PER_HOUR * 6)) {
-        [result appendFormat:@"%d hour%@ ago",
-         hours,
+        [result appendFormat:@"%ld hour%@ ago",
+         (long)hours,
          (hours < 2) ? @"" : @"s"];
     } else {
         NSDateFormatter *formatter = [[NSDateFormatter alloc] init];
